+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.6.2] 2024-08-20
+
+* Implement `equals/hashCode` on `ParkingSpot`
+* Implement `toString` on `AccessLog`, `ApplicationLog`, `Count`, `ParkingSpot`, `WebhookMessage` and `WebhookMessageErrorResponse`
+
 ## [2.6.1] 2024-01-13
 
 * Support for query parameter `subjectId` on `/v2/access-rights`.